+++ packages/AdapterVue3/vite.config.ts
@@ -22,7 +22,8 @@ export default defineConfig({
 
   resolve: {
     alias: {
-      '@': resolve(__dirname, './src')
+      '@': resolve(__dirname, './src'),
+      '@form-renderer/engine': resolve(__dirname, '../Engine/src')
     }
   },