+++ packages/Engine/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @form-renderer/engine
 
+## 1.0.0-alpha.12
+
+### Patch Changes
+
+- update
+
+## 1.0.0-alpha.11
+
+### Patch Changes
+
+- update
+
+## 1.0.0-alpha.10
+
+### Patch Changes
+
+- update
+
+## 1.0.0-alpha.9
+
+### Patch Changes
+
+- update
+
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- Merge remote-tracking branch 'origin/master-A-Fix-12604-xuxin-1202-H5查找多选字段，无法多次搜索选择，第二次搜索选择的数据会把第一次覆盖掉' into master-A
+
+## 1.0.0-alpha.7
+
+### Patch Changes
+
+- update
+
+## 1.0.0-alpha.6
+
+### Patch Changes
+
+- update
+
 ## 1.0.0-alpha.5
 
 ### Patch Changes